% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processor.R
\name{processor}
\alias{processor}
\title{a Processor}
\usage{
processor(
  input_folder.dir = NA,
  output_folder.dir = NA,
  tmp_folder.dir = NA,
  RscriptPathName = NA,
  log_folder.dir = NA,
  cache_folder.dir = NA,
  sync_folder.dir = NA
)
}
\arguments{
\item{parameters.list}{a list containing the parameters. The possible ones are: 'considerAutoLoop' and 'threshold'. 'considerAutoLoop' is a boolean which indicates if the autoloops have to be admitted, while 'threshold' is the minimum value that a probability should have to do not be set to zero, in the transition matrix.}
}
\description{
This is an implementation of the First Order Markov Model (firstOrderMarkovModel) for Process Mining issues.
               This class provides a minimal set of methods to handle with the FOMM model:
             The consturctor admit the following parameters:
parameters.list a list containing possible parameters to tune the model.
}
